Перенос ДТО в контракты (#95)
* Перенос ДТО в контракты

* Fix

* Fix 3
