+++ src/RestaurantService/RestoRate.RestaurantService.Application/UseCases/Restaurants/GetAll/GetAllRestaurantsHandler.cs
@@ -11,7 +11,7 @@ using Mediator;
 
 using Microsoft.Extensions.Logging;
 
-using RestoRate.RestaurantService.Application.DTOs;
+using RestoRate.Contracts.Restaurant.DTOs;
 using RestoRate.RestaurantService.Domain.RestaurantAggregate.Specifications;
 using RestoRate.SharedKernel.Enums;
