+++ src/RestoRate.Contracts/Restaurant/DTOs/AddressDto.cs
@@ -1,4 +1,4 @@
-namespace RestoRate.RestaurantService.Application.DTOs;
+namespace RestoRate.Contracts.Restaurant.DTOs;
 
 public record AddressDto(
     string FullAddress,
