+++ src/RestoRate.Contracts/Restaurant/DTOs/OpenHoursDto.cs
@@ -1,4 +1,4 @@
-namespace RestoRate.RestaurantService.Application.DTOs;
+namespace RestoRate.Contracts.Restaurant.DTOs;
 
 public record OpenHoursDto(
     DayOfWeek DayOfWeek,
