+++ tests/RestaurantService/RestoRate.RestaurantService.IntegrationTests/Helpers/RestaurantTestData.cs
@@ -1,5 +1,5 @@
-using RestoRate.RestaurantService.Application.DTOs;
-using RestoRate.RestaurantService.Application.DTOs.CRUD;
+using RestoRate.Contracts.Restaurant.DTOs;
+using RestoRate.Contracts.Restaurant.DTOs.CRUD;
 using RestoRate.SharedKernel.Enums;
 
 namespace RestoRate.RestaurantService.IntegrationTests.Helpers;
